requirements.txt
pelican
python-dateutil

# gevent==20.6.2
# gunicorn==20.0.4
Pillow==9.0.0
pygments==2.7.4
# pylibmc==1.6.1
# python-dotenv==0.13.0
PyYAML==5.4
# sentry-sdk==0.16.3
# Unipath==1.1
pelican-webassets
webassets
